Fix restart issue
restart issue contains 2 principle backend.

## GLES:

After restart, the context was not destroyed totally, but the swapchain was recreated by using the old context of Device, while the data is out of date.

It's good to re-init the context but only gl backend has this bug. Temperature solution is to destroy the device. It will be recreated after then.

## VK

The surface should be destroyed once we put it background, but at windows platform, the default custom event did not offer the WindowHandler. 

So hack it.